iwlwifi: sdio: turn INTR self-clear optimization on
Starting B-step of 8000 HW family, the INTR self-clear
optimization is supported. In this optimization there
is no need to clear the INT register after reading it.
However, by turning this opt on we need to NOT run in
a loop in iwl_sdio_isr(), otherwise the bus eventually
crashes due to a HW race.

Turn this optimization on to improve throughput on SDIO.

type=feature

Change-Id: I7ff52cb4cfde5d2e9b22493e471bc9b6059fa094
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/9075
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/10361
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: c05554c4791d43fbb5802179f5146425b032e514