iwlwifi: sdio: rename sdio irq handler to avoid duplication
Apparently, the name of the SDIO IRQ thread and handler
in the iwlwifi driver is the same as the one in the
SDIO stack (mmc/core/sdio_irq.c).

Change this duplication to avoid confusion when
debugging with ftrace, even though both are static.

type=cleanup

Change-Id: Iea76ad09e8b7270fae487d59d4f1177385eda2b7
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/9076
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/10231
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 4a12af2b961251d5e88ea6ef9082772561c251c7