iwlwifi: sdio: provide helpers for write{u8,u32}
Passing an u32 / u8 by reference is dangerous since it can
be uninitialized and the compiler won't warn because it will
consider the parameter to be an OUT parameter.
Provide helpers to pass the data to iwl_sdio_ta_write by
value to help the compiler to catch the bugs.

type=cleanup

Change-Id: I8e568f30d7c371df119ca214424b0c9be06b9fe7
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/9133
Reviewed-by: Kaufman, Liad <liad.kaufman@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
Reviewed-by: Peller, EliadX <eliadx.peller@intel.com>
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/10274
x-iwlwifi-stack-dev: 969f398c3722f155dbb8c9795abc839a8b5d9948