iwlwifi: mvm: ROC: Extend the ROC max delay duration & limit ROC duration
When associated to an AP and a ROC event with a long duration is scheduled
the FW may have a hard time scheduling a consecutive time event, since it
has to remain on the connection channel to hear the AP's DTIM.
In addition, when associated and a ROC is requested with a duration
greater than the DTIM interval, the FW will not be able to schedule
the ROC event, since it needs to wake up for the DTIM.

Increasing the "max delay" duration to the DTIM period will allow the FW to
wait until after the DTIM and then schedule the ROC time event.
Limiting the ROC to be less than the DTIM interval will assure that the
time event will be scheduled for at least part of the time (instead of
automatically failing)

Extend the ROC max delay duration to min(dtim_interval * 3, 600TU),
and limit the duration to be less than the DTIM interval.

type=feature

Change-Id: I4d589e50d987663a562352128c09df39640f5704
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/8607
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Peer, Ilan <ilan.peer@intel.com>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
(cherry picked from commit 758cba88d00830b0b00c1cd9f53a30feab592ba8)
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/10351
x-iwlwifi-stack-dev: 63f35fda87c4baf3d462970dfef431602214d829