[BUGFIX] mac80211: fix off-channel mgmt-tx uninitialized variable usage
In the previous commit, I neglected to update the cookie in one code
path: when a mgmt-tx has no real cookie sent to userspace as it doesn't
wait for a response, but is off-channel. The original code used the SKB
pointer as the cookie and always assigned the cookie to the TX SKB in
ieee80211_start_roc_work(), but my change turned this around and made
the code rely on a valid cookie being passed in.

Unfortunately, the off-channel no-wait TX path wasn't assigning one at
all, resulting in an uninitialized stack value being used. This wasn't
handed back to userspace as a cookie (since in the no-wait case there
isn't a cookie), but it was tested for non-zero to distinguish between
mgmt-tx and off-channel.

Fix this by assigning a dummy non-zero cookie unconditionally, and get
rid of a misleading comment and some dead code while at it. I'll clean
up the ACK SKB handling separately later.

type=bugfix
fixes=I341d735dcfa09c1ad4dd84d638639c4321986f49
bug=not-tracked

Change-Id: I1130e02a7f51be05ced46480e0b907f6449194cf
Fixes: 3b79af973cf4 ("mac80211: stop using pointers as userspace cookies")
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/10481
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/13285
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 100ae91d7850b9c51fc33c01269ab5dd00535663