iwlwifi: mvm: Frequency Manager: Change Channel notification to be async
During Connection to an AP the driver may update the FM about frequency
change a few times (depends if the BW > 20 MHZ).

The duration of the API function that is called to notify the FM can take
30-40 ms. This affects the amount of time it takes to connect to the AP.

Change Channel notification to be async, to avoid affect on the duration of
the connection.

type=cleanup

Change-Id: I8215ba81e6a747db3d927de51ee01956c04a116d
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/8606
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/11954
x-iwlwifi-stack-dev: 3309f4fd435bf144f35ace4222c5f97c9cc8ceed