iwlwifi: mvm: add CT-KILL notification
Up to today the driver was notified of the temperature from the FW
and decided whether to enter CT-kill or not.
From now on, the FW will decide when to enter CT-kill and will notify
the driver.
Add support for this notification.

type=feature

Change-Id: I1949b72e0ea32aea3b617307c59e5e455520813f
Signed-off-by: Chaya Rachel Ivgi <chaya.rachel.ivgi@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/18155
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/19723
x-iwlwifi-stack-dev: 4d4d21c2a00b82fe7caa3f2e9897dbdedaff75f7