[BUGFIX] iwlwifi: defconfigs: fix IWLMVM_P2P_UAPSD_STANDALONE pruning
The pruning system for iwlwifi-stack-public-modrename was pruning out
this option, because it was present but not set in the iwlwifi-public
defconfig.  This flag needs to be present, but disabled by default.
To do so, it should not be in the iwlwifi-public defconfig at all.
Instead, it should be added to the publishable-options file.

type=bugfix
bug=not-tracked
fixes=I3cee828052cbec84dbca77225ae42412c489c95f

Change-Id: I3e066b728c75a940cd456f665103b128f0219f03
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/19172
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Reviewed-by: Altman, Avri <avri.altman@intel.com>
Reviewed-by: Mordechai, Dan <dan.mordechai@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/19276
x-iwlwifi-stack-dev: 1f6ccc42ac86d4627288cd3a230d049ff5d3488e