backport: add thermal backports for kernels < 4.3
There were some API changes in the thermal framework in kernel version
4.3 and also earlier in 3.10.  Backport what is needed to support
older kernels.

The 4.3 change is a bit tricky, because it changes the prototypes of
some ops.  The solution for that is to add hook functions that will
intercept the calls from the thermal framework and convert them to
calls that the backported driver provides (namely convert unsigned
longs to ints).

type=feature

Change-Id: I50d172b60d85ebe58246955e6358c895fa6c6c41
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/19637
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/19722
x-iwlwifi-stack-dev: dad95e621b40a146ed2bae5820ca6738c8b62cf4