[BUGFIX] iwlwifi: mvm: Decrease size of the paging download buffer
Currently the driver has 2 buffers for paging:
1. paging db - this contains all of the pages that were in the FW
image, that the driver stores for the FW. This is allocated for each
block separately (not contiguous).
2. download buffer - we need to provide this empty buffer for the
iwl_sdio_load_fw_chunk function to copy the requested pages to the shared
memory. This is one big buffer of contiguous memory whose size is the
size of all the blocks that the fw paging section can contain.

This download buffer size is too big, and causes the allocation to fail
sometimes. Since the driver allocates memory for each block separately,
it is not possible for the FW to request all of the pages in one request
(the FW gives an address and size, so blocks need to be contiguous for
this to happen), therefore the FW is limited to request only one block.

Decrease the size of the paging download buffer to be the size of a
paging block.

type=bugfix
bug=not-tracked
fixes=If9be810bcf633669a851efcce3686b4793a9d7e8

Change-Id: Ie05e7c1e081608c57619934dd0130d283f080a76
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/22910
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Tested-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/23429
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 0e28367b29df271d0f1678036f3842b4ea40b76a