[BUGFIX] iwlwifi: mvm: ROC: cleanup time event info on FW failure
Currently when the FW sends start/stop aux roc time event
notification with an error status, the driver returns an
error value, but does not remove the time event, and does
not notify the stack above that the time event is over.

This causes problems that the stack above assumes we are still
in the middle of a time event, and therefore can block different
events, such as scanning.

On FW failure notification, cleanup the time event parameters and
notify the stack above that the time event is over.

type=bugfix
bug=cq:MWG100254880
fixes=I7c4021a0fbcc73101667487858d13a6eb1ac5918

Change-Id: I691b850f16baee3b5961bb6ef9b19be13c069818
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/23144
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Tested-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/23159
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 3f5337a54825d36642b31960569974a46b6a4d15