[BUGFIX] iwlwifi: backport: restore original ops when unregistering thermal
We overwrite the original ops with our hooks when the thermal_zone
device is registered.  If we unregister and try to register again, the
hooks will be in place of the original ops and we would try to use
them as the actual ops, causing funky things to happen.  To solve
this, copy the ops back to their original places during unregister.

type=bugfix
bug=not-tracked
fixes=unknown

Change-Id: I985617b9090491ad08805e3e4c7bd1c010fcd2e6
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/24023
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/24448
x-iwlwifi-stack-dev: 5f9cb3a69b644c81d36997b1d2c69a00153a4192