[BUGFIX] [CORE18ONLY] iwlwifi: trans_slave: don't take wakelock on STORED_BEACON_NTF
In some cases, we get stored beacon notification after exiting d0i3.
Since we exit d0i3 on .prepare(), taking a wakelock on the following
notification will abort the ongoing suspend.

Since this notification is raised only after exiting d0i3, we can
simply not take the redundant wakelock in this case.

This is Core18-specific and is an equivalent (but not as complete)
fix as commit fe40298d3b2a ("[BUGFIX] iwlwifi: mvm: add a new mvm
reference type for RX data").  The reason for not backporting the
"official" one from master is that it was done after some refactoring
patches and depends on them.  Eliad's patch is proven to work to fix
the bugs found in Core18 (it was cherry-picked and tested in earlier
Android releases).

type=bugfix
fixes=unknown
bug=jira:IMINAN-48991
bug=jira:IMINAN-50278
bug=cq:MWG100260275

Change-Id: Ic0aa3fc4f6fc38986d7410d332c4081c04039608
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-49404
Signed-off-by: Eliad Peller <eliadx.peller@intel.com>
Reviewed-on: https://android.intel.com:443/487575
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/36369
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 8ab02fe736de97d2bff513d9a664a58eead4a573