[BUGFIX] iwlwifi: mvm: Paging: Fix accessing Null pointer during fw dump collection
The firwmare file can come with data that is relevant for paging. This
data is availablet to the firmware upon request, but it stored in the
host's memory. During the firmware init flow, the driver configures the
firmware so that the firwmare knows where is the data.
When paging is used, the variable paging_mem_size is the number of bytes
that are available through paging. This variable is not zeror-ed if the
driver fails to configure the paging in the firmware, but the memory is
freed which is inconsistent.
This inconsistency led to a NULL pointer dereference in the code that
collects the debug data.

Fix this by zero-ing the paging_mem_size variable and NULLify the
relevant pointers, so that the code that collects the debug data will
know that the paging data is not available.

type=bugfix
fixes=Ia25a2b21889a19bb5d7073d79575c031e0a9d190
bug=not-tracked

Change-Id: I6bed25a7349521d61908b0fde4c3fb0baaa7a3a3
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/27588
Reviewed-by: Cohen, Guy <guy.cohen@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/28038
x-iwlwifi-stack-dev: c6e44d31aa21b97c01680fdcca17e4347974ebde