[BUGFIX] iwlwifi: mvm: workaround for a FW bug and pick the beacon transmission's antenna [NOUPSTREAM]
The firmware is currently buggy and can't handle the case
in which we let the firmware choose the antenna to use for
the beacon transmission. As a workaround, always choose the
antenna in the driver until the bug is fixed in the
firmware.

type=bugfix
bug=not-tracked
fixes=unknown

Change-Id: I1e67c7d3f3b3e0a4470cad7a3b26cd2d17dbb6c7
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 3c728f0acfac60adafce3f8e79be81be52f75056