[BUGFIX] iwlwifi: mvm: release static queues on bcast release
A few of the static queues are enabled along with the bcast
STA. Make sure they are removed along with it, rather than
waiting for the mac ctxt release.

This is needed because we sometimes have a STA being removed
and then added again (either with the same sta_id or a
different one). If we wait for the mac ctxt release we will
try to allocate the queues again (as this is currently done
in the STA allocation and not in the MAC init) although
they weren't freed, and even if the sta_id of the STA has
changed.

type=bugfix
fixes=unknown
bug=not-tracked

Change-Id: I368520d7e9ec1258489d45eb6ff238d0094cb241
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/73797
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: lab_WISEJ <unknown@unknown.local.lan>
Reviewed-by: Sharon, Sara <sara.sharon@intel.com>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/78557
x-iwlwifi-stack-dev: 2b8bdfded1b31ebec11574d715cc89e3643c65b0