[BUGFIX] iwlwifi: mvm: fix D3 replay counter value
This fixes a long-standing bug that was introduced when this code
was introduced: cfg80211 passes a pointer, but we treat it as if
it was passing a value. The result is that we pass the pointer to
the firmware, instead of the value. It's not clear how this could
ever have worked, unless the firmware is ignoring this value.

type=bugfix
fixes=unknown
bug=not-tracked

Change-Id: I303d1a37f3f212c47a4ad210a83bf71ce4ae2e80
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/76685
Reviewed-by: Kaufman, Liad <liad.kaufman@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: lab_WISEJ <unknown@unknown.local.lan>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/77848
x-iwlwifi-stack-dev: 017acffc6b5e7d9c0c50f63ff0fac5a6a16eafd6