[BUGFIX]iwlwifi: mvm: fix pending frame counter calculation
In DQA mode the check whether to decrement the pending frames
counter relies on the tid status and not on the txq id.
This may result in an inconsistent state of the pending frames
counter in case frame is queued on a non aggregation queue but
with this TID, and will be followed by a failure to remove the
station and later on SYSASSERT 0x3421 when trying to remove the
MAC.
Such frames are for example bar and qos NDPs.
Fix it by aligning the condition of incrementing the counter
with the condition of decrementing it - rely on TID state for
DQA mode.
Also, avoid internal error like this affecting station removal
for DQA mode - since we can know for sure it is an internal
error.

type=bugfix
fixes=I35b4168139afd9c99d0607ba399ee99320f92f97
bug=jira:WIFILNX-82

Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Change-Id: I82f0612a8eab094604034f970c00913c6d60cf18
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/78541
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: lab_WISEJ <unknown@unknown.local.lan>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Reviewed-by: Kaufman, Liad <liad.kaufman@intel.com>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
(cherry picked from commit 5f91e69dae446a3fe026e8d8c01e3262153021d5)
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/78818
x-iwlwifi-stack-dev: 419bca08f42ab4d58befe5f5af444dce0f32a2e8