[BUGFIX] iwlwifi: pcie: set STATUS_RFKILL immediately after interrupt
Currently, when getting a RFKILL interrupt, the transport enters a flow
in which it stops the device, disables other interrupts, etc. After
stopping the device, the transport resets the hw, and sleeps. During
the sleep, a context switch occurs and host commands are sent by upper
layers (e.g. mvm) to the fw. This is possible since the op_mode layer
and the transport layer hold different mutexes.

Since the STATUS_RFKILL bit isn't set, the transport layer doesn't
recognize that RFKILL was toggled on, and no commands can actually be
sent, so it enqueues the command to the tx queue and sets a timer on
the queue.

After switching context back to stopping the device, STATUS_RFKILL is
set, and then the transport can't send the command to the fw.
This eventually results in a queue hang.

Fix this by setting STATUS_RFKILL immediately when
the interrupt is fired.

type=bugfix
fixes=unknown
bug=not-tracked

Change-Id: I33a3b4b9adea6dac6de1598f6b810ade0f62bedd
Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
Signed-off-by: Golan Ben-Ami <golan.ben.ami@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/79474
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
(cherry picked from commit 8f806a84bee0b79c749c0600c7f14df608dc4c7b)
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/92768
x-iwlwifi-stack-dev: 386b21881a8fa425bb1506d9436f75cc587df2bb