[BUGFIX] iwlwifi: mvm: reset the HW before dumping if HW error is detected
If the hardware is stuck, we can't read any of the memory we need to
dump it, so we end up printing only 0xa5a5a5a5, which is useless.

To solve this, poke the hardware by triggering a reset and re-enabling
the clocks if we detect a HW error.

type=bugfix
fixes=unknown
bug=not-tracked

Change-Id: I87d2b1b534762d7b89cbf62f23e6b1972f393a23
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/113913
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/115911
x-iwlwifi-stack-dev: e4948c9163c1b85e6c48dd9ffbfe874104ba5108