iwlwifi: pcie: delete the Tx queue timer earlier upon firmware crash
When the firmware crashes, the transmit queues can't make
any progress. This is why we stop the counter that monitor
the transmit queues' activity.
The call that notifies the error to the op_mode may take
a bit of time, so stop the timer of the transmit queues
earlier.

type=feature

Change-Id: I8babe8518d9bb4483a961271d94e1ef4a23ecf84
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: db5d85ba08d85dd73a6820221567e6d05569d8c0