[BUGFIX] iwlwifi: pcie: fix 9000-series RF-kill interrupt propagation
A hardware issue on 9000 series devices sometimes causes RF-kill
interrupts to not be propagated to the host properly if ASPM is
enabled. Work around this by setting the right hardware bit to
allow it to interrupt the host for this reason (rfkill).

type=bugfix
fixes=unknown
bug=jira:WIFILNX-798
bug=jira:WIFILNX-688

Change-Id: I459e6eee19bccaa30f0884ececded15ce85b77c5
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/114406
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/117328
x-iwlwifi-stack-dev: 1443be1d960398f0f50405994ef3fd7bea0ab904