[NOUPSTREAM] iwlwifi: mvm: allow vendor command to get SAR profile with the interface down
There's no need to restrict reading the current SAR profile when the
interface is down.  The only drawback is that we currently always set
them back to profile 1 when the interface goes up or when we resume.

type=feature

Change-Id: I48c938698cffa3596d24eda4bdc3f39bbae5885d
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/125038
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/125450
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: d477be5ca6fb7afdbae3edb092e0e2eb58c86459