[BUGFIX] [NOUPSTREAM] mac80211: remove assignment to ktime_to_ns()
In some kernel configurations, specifically < 4.10 with !KTIME_SCALAR,
the ktime_to_ns() symbol is a function, so we cannot assign values to
it.

To fix this, add a check for the kernel version and either use tv64 or
not.

I'm doing this directly here (and not in backports), because this code
is not upstream.  When we upstream this, we need to use always the
non-tv64 version and add a proper backport for it.

type=bugfix
fixes=I66454f2d48b616b2014b93a0023767997183237b
ticket=none

Change-Id: Ie502344e7f932cbcacbbd8352767a9685fbad89f
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/129337
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/130658
x-iwlwifi-stack-dev: 72d59203ddac98719884752be972caf261d46456