[BUGFIX] iwlwifi: mvm: start mac queues when deferred tx frames are purged
Under DQA, when tx is deferred because a queue needs to be allocated,
the mac queue for that TID is stopped until the new stream is added.
If at this point the station that this stream belongs to is removed,
all the deferred tx frames are purged, but the mac queue is not
restarted. As a result, all following tx on this queue will not be
transmitted.

Fix this by starting the relevant mac queues when the deferred tx
frames are purged.

type=bugfix
ticket=jira:WIFILNX-1082
fixes=I3e52044a231a59d63997a0974a4a64ccd0a75cdb

Change-Id: Ie9da653932e176366176fc26e5f6ff43a5af41f1
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/129385
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/130632
x-iwlwifi-stack-dev: cfa43f000b457d4b9810794688186414156b1d73