[BUGFIX] mac80211: call synchronize_net once in the restart flow
Currently the restart flow enables RX back, and then proceeds
to tear down RX and TX aggregations.
The TX aggregation tear down calls synchronize_net(), which
waits for packet receiving to be done.
This is done for every session, while RX processing is already
active, and in some reproductions it takes up to 3 seconds.
Add a call once in the restart_work, before we have traffic
active again, and remove the subsequent calls when tearing
down the aggregation.
This requires to move down the code that turns off the
reconfig flag in order to be able to test it in
_ieee80211_stop_tx_ba_session().

type=bugfix
fixes=unknown
ticket=jira:WIFILNX-1447

Change-Id: Iee9e3514f90366ca9fd53382f2fb5ec270479a88
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/144606
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/147371
x-iwlwifi-stack-dev: 2f0d29bb715229d2abddf08960895081f9ca43b1