iwlwifi: mvm: avoid dumping assert log when device is stopped
We might erroneously get to error dumping code when the
device is already stopped.
In that case the driver will detect a defective value and
will try to reset the HW, assuming it is only a bus issue.
The driver than proceeds with the dumping.
The result has two side effects:
The device won't be stopped again, since the transport
status is already stopped, so the device remains powered
on while it actually should be stopped.
In addition, the dump in that case is completely garbaged
and useless.
Detect and avoid this.
It will also make debugging such issues easier.

type=cleanup
ticket=none

Change-Id: If4516cea643ae55ab87384a71bd654f12babb98d
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/150056
Reviewed-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/150762
x-iwlwifi-stack-dev: 6d5a03c6038a939633dd172654fef7868122a879