[BUGFIX] iwlwifi: mvm: don't warn on multiple packets when opening a queue
When we have TSO enabled, we might end up segmenting it and
queuing multiple packets before the queue is even enabled.  This causes
a warning.

For example, when starting TCP traffic on a non-zero TID, the
first packets may not have DSCP and will be sent on TID 0, while
the actual data packets will be sent on the TID.

To prevent this, simply remove the warning.

type=bugfix
fixes=unknown
ticket=none

Change-Id: Ifd715351c2c5efbd8933d1d2380b51aab014934b
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/153303
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
Tested-by: Coelho, Luciano <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/154120
x-iwlwifi-stack-dev: 9000c20110b17f99ddbeab091128ddd6fdf09609