[BUGFIX] mac80211: don't Tx a deauth frame if the AP forbade Tx
If the driver fails to properly prepare for the channel
switch, mac80211 will disconnect. If the CSA IE had mode
set to 1, it means that the clients are not allowed to send
any Tx on the current channel, and that includes the
deauthentication frame.

Make sure that we don't send the deauthentication frame in
this case.

In iwlwifi, this caused a failure to flush queues since the
firmware already closed the queues after having parsed the
CSA IE. Then mac80211 would wait until the deauthentication
frame would go out (drv_flush(drop=false)) and that would
never happen.

type=bugfix
fixes=unknown
ticket=chrome-os-partner:110347776

Change-Id: I1668d7574ce37a09ae3c27cee444da48b26a6032
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>x-iwlwifi-stack-dev: f60326453fc6ede334a87be3bf35ce9131b62c45